- Bring transaction response type into shared package
